+++ CHANGES.rst
@@ -9,6 +9,7 @@ v9.9.9 (unreleased)
 - chore: bump actions/setup-python version (#1773, #1775)
 - refactor: switch to pyproject.toml (#1753)
 - chore: bump dependencies (#1774, #1777)
+- feat: add md_table helper to render Markdown tables in templates (#1780)
 
 
 v6.2.1 (2026-06-06)